refactor : Update separate the quizCategory component from the createQuiz componenet
