+++ src/pages/CreateQuiz/QuizTextarea/index.tsx
@@ -1,15 +1,15 @@
-import styles from './questionTextarea.module.scss'
+import styles from './quizTextarea.module.scss'
 import { useState } from 'react'
 
-const QuestionTextarea = () => {
+const Quizextarea = () => {
   const [textCount, setTextCount] = useState(0)
   const handleChangeTextarea = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
     setTextCount(e.currentTarget.value.length)
   }
 
   return (
-    <section className={styles.question}>
-      <label className={styles.questionLabel}>
+    <section className={styles.quizArea}>
+      <label className={styles.quizAreaLabel}>
         내용 추가
         <textarea
           className={styles.textarea}
@@ -25,4 +25,4 @@ const QuestionTextarea = () => {
   )
 }
 
-export default QuestionTextarea
+export default Quizextarea
