+++ src/pages/CreateQuiz/index.tsx
@@ -1,41 +1,14 @@
 import styles from './createQuiz.module.scss'
-import cx from 'classnames'
-import QuestionTextarea from './QuestionTextarea'
-import { useState } from 'react'
+import QuizTextarea from './QuizTextarea'
+import QuizCategory from './QuizCategory'
 
 const CreateQuiz = () => {
-  const [disabled, setDisabled] = useState(false)
-  const handleClickDisable = () => {
-    setDisabled((prev) => !prev)
-  }
-
   return (
     <div className={styles.createQuiz}>
       <header className={styles.header}>Header Area</header>
       <main className={styles.main}>
         <div className={styles.display}>Display Area</div>
-
-        {/* 카테고리 */}
-        <section className={styles.category}>
-          <p>카테고리</p>
-          <button
-            type='button'
-            className={cx(`${styles.personality} ${styles.button}`, { [styles.disabled]: disabled })}
-            disabled={!disabled}
-            onClick={handleClickDisable}
-          >
-            인성 질문
-          </button>
-          <button
-            type='button'
-            className={cx(`${styles.technical} ${styles.button}`, { [styles.disabled]: !disabled })}
-            disabled={disabled}
-            onClick={handleClickDisable}
-          >
-            기술 질문
-          </button>
-        </section>
-
+        <QuizCategory />
         {/* 필수질문 지정 */}
         <section className={styles.essential}>
           <label className={styles.essentialLabel}>
@@ -45,8 +18,8 @@ const CreateQuiz = () => {
           </label>
         </section>
 
-        {/* 질문 텍스트 에어리어 */}
-        <QuestionTextarea />
+        {/* 문제 텍스트 에어리어 */}
+        <QuizTextarea />
       </main>
     </div>
   )
